Add Free Throw requests handlers. Add TeamId to foul table. Add migration.

+++ SRC/Application/Shots/Commands/UpdateShot/UpdateShotCommandValidator.cs
@@ -17,7 +17,7 @@ namespace BasketballLeague.Application.Shots.Commands.UpdateShot
             RuleFor(x => x.Quater)
                 .InclusiveBetween(1, 4);
 
-            RuleFor(x => x.ShotType)
+            RuleFor(x => (int)x.ShotType)
                 .InclusiveBetween(1, 7);
 
             RuleFor(x => x.Value)
